Refactor: Move gluten_test_util.h
