+++ cpp-ch/local-engine/tests/CMakeLists.txt
@@ -38,6 +38,8 @@ if(ENABLE_TESTS)
   set(LOCAL_ENGINE_DIR "${ClickHouse_SOURCE_DIR}/utils/extern-local-engine")
 
   add_gtest_sources(local_engine_gtest .)
+  add_headers_and_sources(gtest_utils utils)
+
   if(ENABLE_LOCAL_UDFS)
     file(
       GLOB children CONFIGURE_DEPENDS
@@ -72,8 +74,10 @@ if(ENABLE_TESTS)
 
   add_executable(
     unit_tests_local_engine
-    gluten_test_util.cpp ${local_engine_gtest_sources}
-    ${local_engine_udf_sources} ${local_engine_function_parser_sources})
+    ${local_engine_gtest_sources}
+    ${local_engine_udf_sources}
+    ${local_engine_function_parser_sources}
+    ${gtest_utils_sources})
   target_include_directories(
     unit_tests_local_engine
     PRIVATE ${ClickHouse_SOURCE_DIR}/utils/extern-local_engine
@@ -83,8 +87,7 @@ if(ENABLE_TESTS)
   target_link_libraries(
     unit_tests_local_engine
     PRIVATE gluten_clickhouse_backend_libs clickhouse_parsers loggers
-            ch_contrib::gmock_all ch_contrib::gtest)
-  target_link_libraries(unit_tests_local_engine PRIVATE ch_parquet)
+            ch_contrib::gmock_all ch_contrib::gtest ch_parquet)
 endif()
 
 if(ENABLE_BENCHMARKS)
@@ -92,7 +95,7 @@ if(ENABLE_BENCHMARKS)
                       ${ClickHouse_SOURCE_DIR}/utils/extern-local_engine)
   add_executable(
     benchmark_local_engine
-    gluten_test_util.cpp
+    ${gtest_utils_sources}
     benchmark_local_engine.cpp
     benchmark_parquet_read.cpp
     benchmark_spark_row.cpp
