+++ cpp-ch/local-engine/tests/gtest_parquet_columnindex_bug.cpp
@@ -14,7 +14,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-#include <gluten_test_util.h>
 #include <incbin.h>
 #include <Core/Settings.h>
 #include <Interpreters/Context.h>
@@ -23,6 +22,7 @@
 #include <base/scope_guard.h>
 #include <gtest/gtest.h>
 #include <substrait/plan.pb.h>
+#include <tests/utils/gluten_test_util.h>
 #include <Common/DebugUtils.h>
 #include <Common/GlutenConfig.h>
 #include <Common/QueryContext.h>
