+++ cpp-ch/local-engine/tests/gtest_write_pipeline_mergetree.cpp
@@ -15,7 +15,6 @@
  * limitations under the License.
  */
 
-#include <gluten_test_util.h>
 #include <incbin.h>
 #include <testConfig.h>
 #include <Core/Settings.h>
@@ -37,6 +36,7 @@
 #include <Storages/StorageMergeTree.h>
 #include <gtest/gtest.h>
 #include <substrait/algebra.pb.h>
+#include <tests/utils/gluten_test_util.h>
 #include <Common/BlockTypeUtils.h>
 #include <Common/DebugUtils.h>
 #include <Common/QueryContext.h>
